Move content under pkg folder
